+++ lib/read/cardano-wallet-read.cabal
@@ -48,6 +48,7 @@ library
     Cardano.Read.Ledger.Tx.Outputs
     Cardano.Read.Ledger.Tx.TxId
     Cardano.Read.Ledger.Tx.Cardano
+    Cardano.Read.Ledger.Tx.Certificates
     Cardano.Read.Ledger.Tx.CBOR
     Cardano.Read.Ledger.Tx.CollateralInputs
     Cardano.Wallet.Read
@@ -71,7 +72,6 @@ library
     Cardano.Wallet.Read.Eras.KnownEras
     Cardano.Wallet.Read.Hash
     Cardano.Wallet.Read.Tx
-    Cardano.Wallet.Read.Tx.Certificates
     Cardano.Wallet.Read.Tx.CollateralOutputs
     Cardano.Wallet.Read.Tx.Eras
     Cardano.Wallet.Read.Tx.ExtraSigs
